@@ -1,21 +1,11 @@
 import Vue from 'vue'
 import App from './App.vue'
-import router from './router'
-import store from './store'
 import './registerServiceWorker'
-import BootstrapVue from 'bootstrap-vue'
-import 'bootstrap/dist/css/bootstrap.css'
-import 'bootstrap-vue/dist/bootstrap-vue.css'
-import 'vue-awesome/icons'
-import Icon from 'vue-awesome/components/Icon'
+import router from './router'
 
-Vue.component('v-icon', Icon)
-Vue.use(BootstrapVue)
-window.axios = require('axios')
 Vue.config.productionTip = false
 
 new Vue({
   router,
-  store,
   render: h => h(App)
 }).$mount('#app')
